Add rubocop and rubocop-rails as dependencies
`rubocop -A` was ran after adding the gems to fix all the auto generated
problems, Along with a few top-level comments on classes.

Add `pipeline` command to Makefile. This will also rspec when added. The
aim is to quick run this each time before pushing to origin. Ideally I
would have master be a protected branch with a pipline running these
blocking merge, but it doesn't seem github allows free personal accounts
to have protected branches.