Configure rubocop to ignore generate and doc cops
The spec file is generated automatically so we do not need to cop it. It
is a pointless exercise. Similar for migrations but granted i do edit
these by hand, I still dont think its work policing.

Turn off documentation cops. I like to avoid comments in my code as I
believe it should be self documenting and if that is not possible then
git history can do the rest.