Add the ShortenedUrl model
This model will store the orginial URLs given and the uid will be used
to create the shortened URL.

`uid`: This a case-sensitive alphanumeric string of length 6. Giving us
`1.35^45` possible uids. Will do us for a week or so I reckon.

Next MR will add a root show path that will take this UID as an ID which
will then redirect to the original URL.