Add the redirect path to routes and controller
This is essentially a custom show endpoint, accepting an id which will
be used to find the relevant shortened_url. To keep the link as small as
possible it has been placed at the `/` path. Granted this is not very
RESTful but I think this is a valid purpose to allow for shorter URLs.

Later endpoints touching the ShortenedUrl controller will be RESTful.

If the id does not match a record in the DB it will render the default
Rails 404 page with a :not_found (404) status.

Why 301?
By default rails redirects on 302. This has little change in behaviour
for the user but will for SEO. By setting this to 301 "Moved
Permanently" we tell google that this link isn't important, it is all
about the link we send you to. IE I don't want these shortened links
appearing on googles indexes. TBH for this project it doesn't really
matter since we wont need to think about SEO, but I believe it makes
sense to add now anyway and was little work.