Add association between User and ShortenedUrls
* A User can have many ShortenedUrls
* A ShortenedUrl must have a User

This is half a job atm. Seeds and specs are valid. But the still does
not authenticate that a user is signed in, and so we cannot safely
assume there is a `current_user` to assign the url to. For the time
being to avoid validations the Create actions simple fetches a User,
basically at random.

My next PR will authenticate_user on all but the redirect action. This
will cause quite a few spec changes. Once that is merged and ready we
can easily limit the URLs visible by `current_user.shortened_urls`,
rather than `ShortenedUser.all'

Obviously if this was in production I would have added a feature toggle
and or figured out smaller PRs to enabled effective code review.