Only allow users to see their own URLs
This completes the web pages and makes the sight a little easier to use
since you only see your own links.

The specs for these are not ideal. I clearly have a lot to learn around
how to text the HTML/DOM returned. The glaring flaw here is the index
action. I don't test that the user can only see their own URLs. If I had
a code base to work from I could probably figure it out but the only
references I have are google, which hasn't been much help, and my work
stuff. Which is primarily APIs, so we can easily test the JSON. If I was
part of a team i would definitely be asking for assistance at this point
and would definitely not put this up for code_review.