Validate the original_url contains a protocol
URL validations is biiig task, and is very easy to over restrict. In
this case the only thing we need is for a protocol and then the site
will be able to correct link the user to the url. It is then on them to
ensure they provide a link actually works.